Towers Reworked
- Deleted original fracture and trigger scripts
- Reworked how destruction occurs
- Added Object Destruction and CleanUp scripts
