+++ Level/Assets/Prefabs/Level Design/Seperates/Fractured TowerAlt.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: fccd2301ea257234abd539d609371e1d
-folderAsset: yes
-DefaultImporter:
+guid: 3bd4b0d72a4b80b4aa531344ec10c4ba
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
