+++ Level/Assets/Prefabs/Weapon Prefabs/Cannon Ball.prefab
@@ -49,10 +49,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   rb: {fileID: 3131551523234759030}
-  speed: 50
+  speed: 100
   damage: 25
   destroyTime: 5
-  arcMultiplier: 6
+  arcMultiplier: 4
   barrel: {fileID: 0}
   CannonCamera: {fileID: 0}
 --- !u!64 &5669620086011129602
