+++ Level/Assets/Scripts/CleanUp.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ab953bbf78d164f49aeee46291ae0aa0
+guid: 4356c70297343f64689b2f4a45b0e473
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
