--- Level/Assets/Scripts/Trigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: daba797ca5b0a87478676c3bc512df3b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
