+++ Level/Assets/Scripts/Weapons/Bullet.cs
@@ -51,7 +51,7 @@ public class Bullet : MonoBehaviour
             }
             if(other.CompareTag("Tower"))
             {
-                other.GetComponent<Trigger>().takeDamage(damage);
+                other.GetComponent<FractureObject>().takeDamage(damage);
             }
 
         }
