--- Level/Assets/Unity Store Downloads/Fracture/Documentation.pdf.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 603078d87a2542c42a9c871733d8ca34
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
