--- Level/Assets/Unity Store Downloads/Fracture/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5eb060c80f74335449981133f7bbb3f3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
